// src/platform/plugins/shared/data/common/search/aggs/agg_type.d.ts
type PostFlightRequestFn<TAggConfig> = (resp: estypes.SearchResponse<any>, aggConfigs: IAggConfigs, aggConfig: TAggConfig, searchSource: ISearchSource, inspectorRequestAdapter?: RequestAdapter, abortSignal?: AbortSignal, searchSessionId?: string, disableWarningToasts?: boolean) => Promise<estypes.SearchResponse<any>>;
// ...
export type IAggType = AggType;
export declare class AggType<TAggConfig extends AggConfig = AggConfig, TParam extends AggParamType<TAggConfig> = AggParamType<TAggConfig>> {
// ...
    postFlightRequest: PostFlightRequestFn<TAggConfig>;
// ...
}